The main optimizations are:
- Only clip when necessary
- Reuse the image buffers of intermediate results
- Don't multiply with 1

| Configuration | Before | After | Note |
| --- | --- | --- | ---
| B=0 C=0 | 160ms | 0ms | Pass through is free now
| B=-14 C=0 | 410ms | 210ms |
| B=0 C=22 | 430ms | 260ms |
| B=0 C=-38 | 380ms | 140ms | We can often save `np.clip` when lowering contrast

Tested on a 8468x4320 RGB image.

<details>

Before:
![image](https://github.com/chaiNNer-org/chaiNNer/assets/20878432/007fcc85-a4b3-4a6d-af17-815c409573f8)

After:
![image](https://github.com/chaiNNer-org/chaiNNer/assets/20878432/8cc23eb0-29cf-45f6-8a8a-69a2ca021506)

</details>